+++ numba/tests/test_typedlist.py
@@ -18,12 +18,6 @@ from numba.core.unsafe.refcount import get_refcount
 from numba.experimental import jitclass
 
 
-# global typed-list for testing purposes
-global_typed_list = List.empty_list(int32)
-for i in (1, 2, 3):
-    global_typed_list.append(int32(i))
-
-
 def to_tl(l):
     """ Convert cpython list to typed-list. """
     tl = List.empty_list(int32)
@@ -33,6 +27,7 @@ def to_tl(l):
 
 
 class TestTypedList(MemoryLeakMixin, TestCase):
+
     def test_basic(self):
         l = List.empty_list(int32)
         # len
@@ -504,11 +499,7 @@ class TestTypedList(MemoryLeakMixin, TestCase):
                 self.assertEqual(type(l), list)
 
     def test_catch_global_typed_list(self):
-        @njit()
-        def foo():
-            x = List()
-            for i in global_typed_list:
-                x.append(i)
+        from numba.tests.typedlist_usecases import catch_global
 
         expected_message = ("The use of a ListType[int32] type, assigned to "
                             "variable 'global_typed_list' in globals, is not "
@@ -516,11 +507,12 @@ class TestTypedList(MemoryLeakMixin, TestCase):
                             "constants and there is no known way to compile "
                             "a ListType[int32] type as a constant.")
         with self.assertRaises(TypingError) as raises:
-            foo()
+            njit(catch_global)()
         self.assertIn(
             expected_message,
             str(raises.exception),
         )
+        self.disable_leak_check()
 
     def test_repr(self):
         l = List()
